% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-helpers.R
\name{celsify_temp}
\alias{celsify_temp}
\title{Convert to Celsius}
\usage{
celsify_temp(dat)
}
\arguments{
\item{dat}{Input data frame}
}
\value{
A double vector
}
\description{
Convert to Celsius
}
\examples{
dat <- dplyr::tribble(
~name, ~where, ~temp,
"Adam", "beach", 95,
"Bess", "coast", 91,
"Cora", "seashore", 28,
"Dale", "beach", 85,
"Evan", "seaside", 31)
dat <- localize_beach(dat)
celsify_temp(dat)
}
