% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-helpers.R
\name{localize_beach}
\alias{localize_beach}
\title{localize beach}
\usage{
localize_beach(dat)
}
\arguments{
\item{dat}{Input data frame}
}
\value{
A double vector
}
\description{
localize beach
}
\examples{
dat <- dplyr::tribble(
~name, ~where, ~temp,
"Adam", "beach", 95,
"Bess", "coast", 91,
"Cora", "seashore", 28,
"Dale", "beach", 85,
"Evan", "seaside", 31)
localize_beach(dat)
}
