% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-helpers.R
\name{outfile_path}
\alias{outfile_path}
\title{Outfile Path}
\usage{
outfile_path(infile, time = Sys.time())
}
\arguments{
\item{infile}{input file name}

\item{time}{use can provide a time, but "now is the default.}
}
\value{
A character vector
}
\description{
Outfile Path
}
\examples{
infile <- "raw.csv"
outfile_path(infile)
}
